editar deporte working now!!

+++ crud-test/crud-melaku/src/main/java/com/example/crudmelaku/DeporteRepository.java
@@ -2,7 +2,7 @@ package com.example.crudmelaku;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeporteRepository extends JpaRepository<Deporte, Long> {
+public interface DeporteRepository extends JpaRepository<Deporte, Integer> {
     public Deporte findById(int id);
 
     public int deleteById(int id);
